update
added input validation in payment page.
seatplan, payPage, paySuccessPage cannot be accessed when user is not logged in.
added checkers if there's a user session indicating a user is logged in (if not will be redirected to loginpage).
changed booking_seats db.
seatplan still not working.
reservation still not working.